[02.12-안승우] 작업해요/판매작품 수정

+++ app/src/main/java/io/sinzak/android/ui/main/profile/art/SaleViewModel.kt
@@ -5,6 +5,7 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import io.sinzak.android.enums.Page
 import io.sinzak.android.model.market.ProductDetailModel
 import io.sinzak.android.model.profile.ProfileModel
+import io.sinzak.android.remote.dataclass.product.Product
 import io.sinzak.android.ui.base.BaseViewModel
 import io.sinzak.android.ui.main.profile.art.adapter.SaleWorkAdapter
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -23,8 +24,8 @@ class SaleViewModel @Inject constructor(
      */
     val isCompleteList = MutableStateFlow(false)
 
-    val adapter = SaleWorkAdapter(productModel::endTrade){ a ->
-        productModel.loadProduct(a.id!!.toInt())
+    val adapter = SaleWorkAdapter(productModel::endTrade){ a->
+        productModel.loadProduct(a.id!!)
         navigation.changePage(Page.ART_DETAIL)
     }.apply {
         model.productList.onEach {
@@ -40,8 +41,12 @@ class SaleViewModel @Inject constructor(
         }.launchIn(viewModelScope)
     }
 
+
     fun setIsComplete(status : Boolean)
     {
         isCompleteList.value = status
     }
+
+
+
 }
